2022 day11 refactor

+++ year2022/go22/cmd/day11/day11_test.go
@@ -16,7 +16,7 @@ func Test_correctMonkeyBusiness(t *testing.T) {
 	monkeyBusiness := findMonkeyBusiness(monkeys, 20, true)
 
 	// Assert
-	if diff := cmp.Diff(monkeyBusiness, int64(10605)); diff != "" {
+	if diff := cmp.Diff(monkeyBusiness, 10605); diff != "" {
 		t.Error(diff)
 	}
 }
@@ -30,7 +30,7 @@ func Test_correctMonkeyBusinessNoWorry(t *testing.T) {
 	monkeyBusiness := findMonkeyBusiness(monkeys, 10_000, false)
 
 	// Assert
-	if diff := cmp.Diff(monkeyBusiness, int64(2713310158)); diff != "" {
+	if diff := cmp.Diff(monkeyBusiness, 2713310158); diff != "" {
 		t.Error(diff)
 	}
 }
